+++ lib/beyond_canvas/models/shop.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+module BeyondCanvas
+  module Models
+    module Shop # :nodoc:
+      extend ActiveSupport::Concern
+      include BeyondCanvas::Models::Authentication
+      include BeyondCanvas::Models::Utils
+
+      included do
+        attr_accessor :api_url, :return_url
+
+        ##############################################################################
+        # Validations
+        ##############################################################################
+
+        # Callback url params
+
+        validates :api_url,
+                  presence: true,
+                  on: :create
+        validates :return_url,
+                  presence: true,
+                  on: :create
+      end
+    end
+  end
+end